#710 renamed waste.lua to waste_flow.lua
